Refactor & fix page editor integration
