+++ src/components/event_maker/pages/CommonInputs.jsx
