+++ src/components/event_maker/pages/components/Image.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+export function Image({
+    src = "https://via.placeholder.com/150",
+    alt = "Placeholder",
+    className = "",
+}) {
+    return <img src={src} alt={alt} className={`rounded shadow ${className}`} />;
+}
+
+Image.propTypes = {
+    src: PropTypes.string,
+    alt: PropTypes.string,
+    className: PropTypes.string,
+};
