+++ src/components/event_maker/pages/components/index.js
@@ -1,0 +1,15 @@
+import { Title } from "./Title";
+import { Image } from "./Image";
+import { Button } from "./Button";
+import { Text } from "./Text";
+import { Alert } from "./Alert";
+
+export const componentList = [
+  { name: "Title", title: "Title Component", component: Title },
+  { name: "Image", title: "Image Component", component: Image },
+  { name: "Button", title: "Button Component", component: Button },
+  { name: "Text", title: "Text Component", component: Text },
+  { name: "Alert", title: "Alert Component", component: Alert },
+];
+
+export { Title, Image, Button, Text, Alert };
