+++ src/contexts/UsersContext.jsx
@@ -8,6 +8,7 @@ const UsersContext = createContext();
 export const UsersProvider = ({ children }) => {
     const usersBaseUrl = `${baseUrl}/users`;
     const usersRolesUrl = `${baseUrl}/users/roles/users`;
+    const { keycloak, initialized } = useKeycloak();
 
     const [users, setUsers] = useState([]);
     const [usersGroupedByRole, setUsersGroupedByRole] = useState({});
@@ -15,10 +16,16 @@ export const UsersProvider = ({ children }) => {
     const [error, setError] = useState(null);
 
     const fetchUsers = async () => {
+        if (!initialized || !keycloak?.authenticated) {
+            console.log("Keycloak not initialized or user not authenticated");
+            return;
+        }
+
         setIsLoading(true);
         setError(null);
         try {
-            const response = await axios.get(usersRolesUrl);
+            console.log("Fetching users with auth status:", keycloak.authenticated);
+            const response = await axiosWithAuth(keycloak).get(usersRolesUrl);
             console.log("Users grouped by role fetched successfully:", response.data);
 
             setUsersGroupedByRole(response.data);
@@ -160,8 +167,7 @@ export const UsersProvider = ({ children }) => {
                 error
             );
             setError(
-                `Failed to ${
-                    shouldBeBanned ? "ban" : "unban"
+                `Failed to ${shouldBeBanned ? "ban" : "unban"
                 } user. Please try again later.`
             );
             throw error;
@@ -179,8 +185,10 @@ export const UsersProvider = ({ children }) => {
     };
 
     useEffect(() => {
-        fetchUsers();
-    }, []);
+        if (initialized && keycloak?.authenticated) {
+            fetchUsers();
+        }
+    }, [initialized, keycloak?.authenticated]);
 
     // Memoize the value object to prevent unnecessary re-renders
     const contextValue = useMemo(
